requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
